+++ backend/auth/passportutil.js
@@ -6,7 +6,7 @@ const LocalStrategy = require("passport-local").Strategy;
 const JWTstrategy = require("passport-jwt").Strategy;
 const ExtractJWT = require("passport-jwt").ExtractJwt;
 
-const UserModel = require("../model");
+const UserModel = require("../models/UserModel");
 
 passport.use(
   new JWTstrategy(
