+++ backend/index.js
@@ -5,7 +5,7 @@ const bodyParser = require("body-parser");
 
 // const UserModel = require("./models/model");
 
-mongoose.connect("mongodb://127.0.0.1:27017/passport-jwt", {
+mongoose.connect("mongodb://127.0.0.1:27017/MAW-DB", {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
@@ -13,36 +13,21 @@ mongoose.set("useCreateIndex", true);
 mongoose.connection.on("error", (error) => console.log(error));
 mongoose.Promise = global.Promise;
 
-require("./models/auth/auth");
+require("./auth/passportutil");
 
-const routes = require("./routes/routes");
-const secureRoute = require("./routes/secure-routes");
+const authRoutes = require("./routes/AuthRoutes");
+const userRoutes = require("./routes/UserRoutes");
 
 const app = express();
 
 app.use(bodyParser.urlencoded({ extended: false }));
 
-app.use("/", routes);
-
-// Plug in the JWT strategy as a middleware so only verified users can access this route.
-app.use("/user", passport.authenticate("jwt", { session: false }), secureRoute);
-
-// require('.models/auth/auth');
-
-// const authRoutes = require('./routes/authroutes');
-// const userRoutes = require('./routes/userroutes');
-
-// const app = express();
-
-// app.use(bodyParser.urlencoded({ extended: false }));
-
-// app.use('/auth', authRoutes); // authentication routes are not JWT protected
-// app.use('/user', passport.authenticate('jwt', { session: false }), userRoutes); // all user routes are JWT protected
+app.use("/auth", authRoutes); // authentication routes are not JWT protected
+app.use("/user", passport.authenticate("jwt", { session: false }), userRoutes); // all user routes are JWT protected
 
 // Handle errors.
-app.use((err, req, res /* next */) => {
+app.use((err, req, res, next) => {
   res.status(err.status || 500);
-  console.log("ERRORRRR");
   res.json({ error: err });
 });
