+++ backend/models/UserModel.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 
-mongoose.connect("mongodb://127.0.0.1:27017/test");
+mongoose.connect("mongodb://127.0.0.1:27017/MAW-DB");
 // const uniqueValidator = require('mongoose-unique-validator');
 const bcrypt = require("bcrypt");
