--- backend/routes/secure-routes.js
@@ -1,13 +1,0 @@
-const express = require("express");
-
-const router = express.Router();
-
-router.get("/profile", (req, res /* next */) => {
-  res.json({
-    message: "You made it to the secure route",
-    user: req.user,
-    token: req.query.secret_token,
-  });
-});
-
-module.exports = router;
